requirements.txt
requests
progressbar2
aiohttp
sanic
sanic-scheduler